Break Off Update #1
This is the first in a series of updates, aiming to make the blocks
break apart on a timer. So far, in this update, the blocks simply are
destroyed, using two scripts, "DestroyTimer" and "DestroyOnTimer".
"DestroyTimer" is simply to hold the timer in a GameObject.
"DestroyOnTimer" accesses the timer and checks whether it matches the
object's Z position before being destroyed.

In "DestroyOnTimer", the code for the destroy event can be changed later
in the series of updates to specify movement when being destroyed,
depending on what object is running the script. Here, tags, names or
varying scripts can be used, according to what works best.